Reorganized modules
* Moved “device/state.rs” to “lib.rs”
* Moved “device/target.rs” to “target.rs”
* Moved “render/state.rs” to “block.rs"
