--- src/device/mod.rs
@@ -1,4 +1,0 @@
-//! State blocks for device.
-
-pub mod state;
-pub mod target;
