--- src/render/mod.rs
@@ -1,3 +1,0 @@
-//! State blocks for rendering.
-
-pub mod state;
